+++ src/main/java/com/dotwait/integer/ModularExponentiation.java
@@ -13,9 +13,9 @@ public class ModularExponentiation {
      * @param m 模数，非负整数
      * @return 取模结果
      */
-    public static int modular(int b, int n, int m){
-        int modResult = 1;
-        int power = b;
+    public static long modular(long b, long n, long m){
+        long modResult = 1;
+        long power = b;
         int nLen = IntegerExpansion.constructBinaryExpansion(n).length();
         for (int i=0;i<nLen;i++){
             if (((n >> i) & 1) == 1){
